Web_Scrapping.py
Web Scrapping methodologies to support Natural Language Processing for N Gram model of LLM. The idea is to identify the positive comments on the business growth of a stock. It could recommend the customer whether the Stock ticker for Financial institutions to be a potential Buy option. The discussion on the ticker would be recommended once the Scrapping content is send as an input to the model along with our dictionary of significant information related to changes in the industry or Organizational growth factors  impacting suggested outcome to be one of the two possible classification labels(Buy,Sell).